Refactor dto package to model
